% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{data.general_a}
\alias{data.general_a}
\title{Database general of serie a 2020}
\usage{
data.general_a()
}
\value{
table with all the matches played in 2020
}
\description{
Database general of serie a 2020
}
\examples{
data.general_a()
}
