% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PB.R
\name{PB}
\alias{PB}
\title{Web scraping of PB championship}
\usage{
PB()
}
\value{
two tables with the matches played and future matches
}
\description{
Open Firefox browser in GE website simulating a user
collecting the local of match, scoreboard and the team names
}
\examples{
PB()
}
