% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serie_a.R
\name{serie_a}
\alias{serie_a}
\title{Web scraping of serie a}
\usage{
serie_a()
}
\value{
two tables with the matches played and future matches
}
\description{
Open Firefox browser in GE website simulating a user
collecting the local of match, scoreboard and the team names
}
\examples{
serie_a()
}
