% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serie_b.R
\name{serie_b}
\alias{serie_b}
\title{Web scraping of serie b}
\usage{
serie_b()
}
\value{
two tables with the matches played and future matches
}
\description{
Open Firefox browser in GE website simulating a user
collecting the local of match, scoreboard and the team names
}
\examples{
serie_b()
}
