% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_env.R
\name{set_env}
\alias{set_env}
\title{Settings}
\usage{
set_env()
}
\value{
set the configuration needed for the libs used on package.
}
\description{
Creates a virtual environment with needed libs
for the functions of soccerdatas package.
}
\examples{
set_env()
}
