% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_ve.R
\name{set_ve}
\alias{set_ve}
\title{Settings}
\usage{
set_ve()
}
\value{
set the configuration needed for the libs used on package.
}
\description{
Creates a virtual environment with needly libs
for the functions of soccerdatas package.
}
\examples{
set_ve()
}
